+++ setup.py
@@ -46,7 +46,7 @@ setup(
         'pytest-runner~=4.0',
     ],
     install_requires=[
-        'anonymous@git+https://github.com/FINNGEN/finngen-common-data-model#egg=finngen-common-data-model-0.0.1',
+        'finngen_common_data_model@git+https://github.com/FINNGEN/finngen-common-data-model.git#egg=finngen_common_data_model',
         'Flask>=0.12',
         'Flask-Compress~=1.4',
         'Flask-Login~=0.4',
